requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.3
cryptography==44.0.1
dj-database-url==3.0.1
Django==5.2.6
django-allauth==65.4.1
django-cors-headers==4.8.0
django-environ==0.12.0
django-filter==25.1
django-ipware==7.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.10
google-auth==2.40.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
lucide==1.1.3
msgpack==1.1.0
optional-django==0.3.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
react==4.3.0
redis==5.2.1
requests==2.32.5
rest-framework-simplejwt==0.0.2
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.10.0
